LayoutTests/http/tests/security/block-top-level-navigation-via-redirect-by-third-party-iframes-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to initiate navigation for frame with URL 'http://127.0.0.1:8000/security/block-top-level-navigation-via-redirect-by-third-party-iframes.html' from frame with URL 'http://localhost:8000/security/resources/navigate-top-level-frame-to-failure-page-via-redirect.html'. The frame attempting navigation of the top-level window is cross-origin or untrusted and the user has never interacted with the frame.
Test blocking of suspicious top-level navigations by a third-party iframe (same-site but redirects to a different site)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS All navigations by subframes have been blocked
PASS successfullyParsed is true

TEST COMPLETE
